BAM-164: release from an input tag (#22)
**Associated JIRA tasks**

BAM-161

**What does the change do?**

For a new project, we may want to release an SDK with an
alpha(prerelease) version of our specification to our customers to
gather early feedback. The alpha version of the specification might
still be under review and reside in a task or feature branch.
Additionally, for developers, it can be beneficial to release a
prerelease version of the SDK. This allows them to begin working on a
specification that hasn’t been fully finalized, especially in situations
requiring urgent progress.

**What has been changed**

To be able to release the such sdk from an alpha version spec, the
action from the sdk repo need to checkout the protobuf spec using a
manual input tag.
Also included new spec to be published for user to try

**Tests**
https://github.com/KodyPay/kody-clientsdk-java/actions/runs/12832138884
was failed due to running from task branch. It should pass once merged.